refactor(react-bloc): ♻️ move context creation to its own file
