+++ packages/react-bloc/src/lib/hooks/useRepository.ts
@@ -1,5 +1,5 @@
+import { AnyClassType } from '../context';
 import { useProvider } from './useProvider';
-import { AnyClassType } from '../components';
 
 export const useRepository = <Repository extends AnyClassType>(
   repository: Repository
