+++ packages/react-bloc/src/lib/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './hooks';
 export * from './util';
+export * from './context';
